requirements.txt
numpy==1.24.3
pandas==2.2.1
PyYAML==6.0.1
torch==2.0.0
torchinfo==1.8.0
